make review requested changes
