+++ index.html
@@ -81,7 +81,6 @@
         openmct.install(openmct.plugins.Filters(['table', 'telemetry.plot.overlay']));
         openmct.install(openmct.plugins.ObjectMigration());
         openmct.install(openmct.plugins.ClearData(['table', 'telemetry.plot.overlay', 'telemetry.plot.stacked']));
-        openmct.install(openmct.plugins.WebPage());
         openmct.start();
     </script>
 </html>
