+++ src/plugins/webPage/components/webPage.vue
@@ -11,17 +11,6 @@ export default {
         return {
             currentDomainObject: this.domainObject
         }
-    },
-    methods: {
-        updateDomainObject(newDomainObject) {
-            this.currentDomainObject = newDomainObject;
-        }
-    },
-    mounted() {
-        this.unobserve = this.openmct.objects.observe(this.domainObject, '*', this.updateDomainObject);
-    },
-    destroyed() {
-        this.unobserve();
     }
 }
 </script>
